feat: split management routers
